Extract Factories to dedicated files
